Lvm: Fixed recaching of lvm-filter
The main issue was that the Map for the cache accepted
Collection<String> as a key. From different sources, the same "set"
of strings was added as a Set and as a List. At least the .equals
of Lists only work for Lists and Set.equals also only work for Sets.

That means that Listor removed instance of the "same" key, but the
other instance was still found to prevent unnecessary external command
calls.

After this patch only Sets are allowed, not arbitrary Collections.