Openflex: Allow Provider for Satellite and Combined type
Basically there is no reason why a normal satellite should not act
as an openflex-target manager.
Although the user will not be able to put an nvme-initiator resource
on the same node obviously.
So in most cases this might be considered as wasting a satellite.
However, in other setups (like k8s), it might be easier to waste
a satellite than make the controller capable of running (even if only
fake) satellites (in terms of privileges and such).