StltExtToolsChecker: Replaced ExtCmd with native Files.readAllBytes
Reading the content of /proc/modules is not done via Files.readAllBytes
instead of calling `cat /proc/mdoules` as external command and parsing
its output.

As now the trace-logs will not include the output of `cat /proc/modules`
this patch also adds some extra tracelogging if Linstor can omit the
modprobe when finding the requested module in the parsed /proc/modules.