SharedSpMgr: Second version
CAUTION: This version is still broken / wip!
See details below.

The SharedStorPoolManager still have utility methods to "lock by
resource", but that is now mapped to "lock by node and a collection
of SharedStorPoolNames". This way, locks are now acquired by nodes
and one node can only have one set of locked sharedStorPoolNames.

That makes releasing the locks also easier as the caller only have
to specify the node of which the locks should be released.

On the satellite side, the DeviceManager now scans the dispatched
resources and snapshots for sharedStorPools that need locking,
sends the controller a locking-request and blocks until the
controller grants the locks.
This is done in a phase after the objects (rscs, snapshots, node, ..)
were already applied, which means that in the meantime new changes
are queued for the next DeviceManager run.

In every case the DeviceManager sends a "DeviceManagerRunComplete"
message to the controller which will automatically release all
previously acquired locks (if any).
This also triggers the controller to send new messages grating
requested locks to other satellites.

wip status: Although this commit could be enough to make sure that
only one satellite can concurrently access a shared storagePool
(although more testing is required), this commit DOES NOT make sure
that only one satellite creates a resource based on a shared LV.