SharedSp: Added external_locking to SP
All storage pools now have an additional boolean "externalLocking"
which is also persisted to the DB.
This boolean only makes sense for shared storage pools (i.e.
SPs with a non-default --shared-space name).

If this boolean is true, Linstor's new SharedStorPoolManager will
consider this shared storage pool as non-shared, i.e. no locks
required to operate on that storage pool.

The intention for this new boolean is the usage of external locking
mechanisms as cLVM.