AutoTiebreaker: fixed race condition
When the autotiebreaker takes over a diskful resource as a diskless
(with toggle disk), the toggle disk ApiCallHandler sends out an
updateSatellite. As the requiresUpdateFlux was also set to true,
a second updateSatellite is sent.

The toggle-disk (first updateStlt) gets the expected answer. However,
as the ctrl continues with the multi-stepped toggle-disk-flux, the
satellite already processes the second (wrongly sent) update.
When the controller finishes, a third updateStlt is sent (as expected).

That one, however, is answered by the second updateStlt with still the old
lvm volume data instead of the in the meantime updated diskless volume
data.
That causes the exception during merging the layers.

This is still quite fragile, so we should think about a more stable fix
where "unneccessary" updateSatellites are handled correctly.