Tests: Fixed security setting for tests
Disabling linstors security was only done in the tearDownClass
method of GenericDbBase. That means, if the first class using that
commong super-class could failed because of the active security
setting.

This patch moves the disabling code to the setupBeforeClass
method so also the first class has to explicitly enable security
if required.