ExtCmdFactoryStlt: Prohibits new sharedSP-calls
ExtCmdFactoryStlt instances can be configured that the ExtCmd they
provide will be used for external calls possibly modifying shared
storage.

Whenever the satellite loses connection to the controller, all
registered ExtCmdFactoryStlts that were marked for shared storage
usage will be instructed to kill all ongoing ExtCmd (if possible)
and to prevent any further execution of ExtCmds.

This is implemented based on the DeviceManger's decision if
it got all locks granted it requested previously. If the controller
goes offline, the DeviceManager pretends that no lock is granted
anymore, but some might still be requested.

If the DeviceManager run finishes (regardless if with or without
an exception, like ExtCmdConditionNotFullfilledException), the
DeviceManager also clears its requested locks, which means the
next DeviceManager run will be automatically unlocked.

The next DeviceManager run should only occur when the controller
established a new connection - and the controller can be asked
for the locks again as usual.