DevMgr: Inverted accesses to svcCondFlag
The methods drbdStateAvailable and fullSyncApplied setting the
svcCondFlag to true as well as the catch(SvcCondException) setting
svcCondFlag to false indicate the usage of the conditionFlag as
"true -> go, false -> everything should stop"

However, the usage within ifs and loops was the exact oposite.

That means, the first devMgr-run that had the svcCondFlag set to
true (by fullSyncApplied or drbdStateAvailable) threw immediately
a SvcCondException, this set the svcCondFlag to false, which was
the sign for the next runs to go as "intended".