DevMgr: Attempt to fix deadlock when requesting shared locks
Requesting shared locks did hold writelocks of reconf, nodes, etc..
The FullSync (which triggers the first DevMgr) also tried to
query FreeSpace after it triggered DevMgr, whereas the FreeSpaces
also want readlocks -> deadlock

The DevMgr now locks the objects for calculating the required
shared locks, but releases the reconf, nodes, ... locks before
requesting (and waiting) for the shared locks from the controller.

As there is a timeframe between "request shared locks" and
"execution requiring those shared locks" where the DevMgr had to
release the reconf, .. locks, there is a second check in the actual
execution phase re-verifying if the before calcualted shared locks
are "enough". If not, an exception is thrown and the locks are
released. Then it's the controllers turn to decide to retry or
what to do.