LayerData: Added nullchecks
404f9819c8ee4ace14abb137aa93fdbc0f61e77d added the functionality
that the LayerDataHelper can now determine the used StoragePools
without needing the volumes to be created (otherwise we would end
up in a chicken/egg problem as the rsc-layerdata have to be created
before volumes can exist, but the rsc-layerdata also need to know
which storage pools will be used, which is usually obtained
through volumes...).

However, the storPoolResolveHelper.resolveStorPool cannot access
any resource- or volume-properties (where the user-defined SP would
be placed) as those objects (rsc, vlm,...) simply does not exist
currently. Therefore using the "resolveStorPool" is only needed
if the payload does not contain a storPoolName for all StorageVlms