Autoplace: Fix incorrectly logged rsc-count
When the Autoplace-API calculates that there is 0 additional
resources to auto-place, the log-message still counted the
diskless resource. This can be confusing as the user get an ERROR
instead of a WARNING if the additional count goes below 0.
Seeing a WARNING and the log of having i.e. 3 nodes while trying
to autoplace 2 is confusing.