StorConfReader: Fixed NPE
If the user specifies a relative path like "tmp" the StorConfReader
runs into a NullPointerException when checking
Files.exists(path.getParent()) as .getParent returns null.

Moving the .isAbsolute check infront of the .getParent does solve
this issue, but the user should still be aware where the storage-pool
is created, and therefore linstor does not try to create all missing
parents, only the bottom most if needed.