node-delete: Check that node delete wouldn't fail because of resources
If a resource of the node is in use or would be the last diskful node,
the node delete would fail and leave a half deleted node in place.
Also it may happen that someone would remove a used node by mistake and
this prevents any catastrophes that would need cleanup.