DBLoad,DrbdExtMd: ignore extMdProp when loading from DB
If a resource was created with internal metadata, and the
external-metadata property was set after the mentioned resource-
creation, restarting the controller will fail as the DatabaseLoader
will also use the now set property to determine whether or not
the DRBD resource should have external metadata.
As the property is now set, the DRBD resource will respond that
it indeed needs a .meta child object. As there was no such child
loaded from the database, the controller attempts to create this
child, which fails as the numberpools are not yet initialzed.

This patch fixes this issue by skiping the property check during
DB-loading as during this time no "completely new" object should
be created anyways (as that would also lead to new INSERT statements
which is generally a very bad idea during DB-loading...)