Autoplace+Tiebreaker: Fixed incorrect storpoolFilter
This commit is quite similar to
c53d08c6ed999eaf03574876e14b6172e278aacd
but now we also need to ignore the second filter-source.

Copied from the java comments from the method finding a storage pool
for the tiebreaker after the decision that we do need a tiebreaker:

There are two possibilities:
1) we just autoplaced at least 1 diskful resource so we reached
the 2 diskful + 0 diskless condition that we now place a tiebreaker.
That means that if the ctx.selectFilter.getStorPoolNamesList is not
null it will contain the name(s) of disk*ful* storage pool(s). We
need to ignore those as otherwise we will not be able to find a
diskless storage pool for our tiebreaker

2) we just autoplaced at least 1 diskless resource. If that is the
case, the condition of 2 diskful + *0* diskless cannot be fulfilled,
so we cannot be here :)