Drbd: Fix ImplError when deleting a VlmDfn
When deleting only a VlmDfn instead of the whole resource, the
DrbdLayer goes the "adjust" path instead of the "delete" path since
the delete-path would delete the entire local DRBD-resource instead
only the one volume.

However, on the adjust-path nothing sets the drbdVlmData.exists to
false which triggers an ImplError in the DeviceHandlerImpl when
checking if all layers successfully have set that boolean.