Exos: Added _EVN properties for IP, Username and Pw
If the actual property is set, it will be taken
Otherwise, if the property *Env is set, its value is used to load
the corresponding environment variable.

Additionally, Username and Pw can now also be set above enclosure
level. However, the actual properties is always prioritized above
the *Env version