MakeRscAvail: First version
Has no support for shared-sp (not merged to master yet).

If the resource already exists and a layer-list is given, the
API checks if the existing resource matches the requested layer-list.

If the resource does not exist, the API
* if DRBD is in the layer-list (or layer-list is empty)
  * creates a diskless DRBD resource if at least one diskful peer
  exists
  * creates a diskful DRBD resource if no diskful peer exists, or
  if 'diskful' was set to true
* if NVMe or OpenFlex is in the layer-list
  * creates an initiator if a target already exists
  * creates a target otherwise
* creates the resource just as usually.

If a storage pool is needed, the autoplacer selects one on the
given node.