Fixed dfltAddrSupplier algorithm in NetInterfaceTestFactory
This one had a quite subtle bug. The intention was to generate
IP addresses for 255.255.255.0 subnets, but to skip the
network address itself (.0) and the broadcast address (.255).

The way it was implemented, it worked in the range from
10.0.0.1 to 10.0.1.254, successfully skipping the addresses
10.0.0.254 and 10.0.1.0, but the offset of the IP addresses
to be skipped begins to slip at higher ranges, producing
the sequences
10.0.1.254
10.0.1.0
10.0.2.2
and later
10.0.2.254
10.0.2.0
10.0.2.1
10.0.3.3
10.0.3.4
etc.

The fix should handle all IP addresses up to 10.0.255.254 correctly,
but does not detect going out of range, which seems to be alright,
as it appears to be used only for a more limited number of test cases.