RscList: Added shared_name
linstor resource list will now additionally contain a shared_name
field for each resource. That field is the SharedStoragePoolName
of the storage pool of the first data volume of the current resource.

The intention for this new field is to find out which set of
resources are shared with each other.

This field is not perfect, as (in theory) one could still create
a resource with 2 (or more) volumes all being in differently
shared storage pools. If this happens, please contact LINBIT.