CI: use docker safe refname for virter test version
CI uses the tests version to determine which docker image to use. This
must be a "docker-safe" tag, i.e. no slashes allowed. The image was already
pulled correctly, but the forwarded version in the test scripts itself
remained the raw branch name. This commit replaces this with the docker-safe
ref.