ToggleDisk: Another fix for resolving storage pools
1e85e0d458247895d50536c27fc16b4cf6d96870
Toggle-disk: fixed bugs with shared SP
.. tried to fix a bug where the storage pool needs to be resolved.
However, in the case of creating a resource, the resolve step is
not necessarily needed  (i.e. if all storage pools are referenced
by the LayerPayload).

The previous fix actually changed the behavior of creating diskless
resources if i.e. the RG had the property StorPoolName set to a
storage pool that existed. Before 1e85e0d4 the diskless creation
ignored that property but afterwards tried to use it which failed
as a diskful storage pool cannot be used as a diskless one.

This patch should work for both scenarios