containers: get pkgs in first stage
Fetching needed pkg in the first stage has the advantage that we don't
need to do an architecture split. As long as there is something that is
like a Centos (and has this name) then we can use it to fetch packages.
We don't need to handle different repos/keys and such. We could even
make the name an ARG, but for now keep it as is, the build system
replaces this name with one that already contains the build
dependencies.

Later we can then just copy these packages to the final stage and use
them. In general this could also be used to cache them and only fetch
them in the first stage if they don't exist.

The only disadvantage and why this was implemented differently in the
first place is that it makes dependency resolving harder as one needs to
fetch the packages and all their dependencies that don't exist in the
final stage. This is trivial here and even if we would fail, we would
notice that early on the next build, so no harm in that. Let alone it
saves a ton of ugly shell scripting, and dealing with keys and repos.

In the future this will also make it easier to replace Centos as pkg
provider.