RscApplied: Only merge Satellite props-namespace
Instead of merging all properties from a resource and its volumes
the controller now only merges the "Satellite/" namespace.

That is to prevent the scenario where the satellite did not received
the most current version and might therefore delete / restore a property
which was added / deleted in the meantime by the controller intentionally.

This means the "Satellite/" namespace should be considered as read-only
by the controller!