containers,satellite: add ext/xfs tools
I was pretty sure we have them, but we don't. We can not just install
them in the final stage as this is based on an UBI image and they don't
provide them. So use the old trick and borrow them (and deps) from an
earlier Centos based stage.

yumdownloader also fetches i686 packages, just delete these. This helps
on x86_64, and does not hurt on the other archs.