Add migration fixing duplicate layer-data
Due to a bug in 1.12.0, resources that were restored from snapshots
created too many layerdata. The resource works fine afterwards,
but the unused layerdata are not properly deleted when the resource
should be deleted, thus causing constraint violations when trying
to finally delete the resource itself.

This patch removes those unused entries in the layer tables