docker: create error when requested package is missing
yum by default does not report an error if at least one package name
matched in a install list with multiple packages. this meant that
prior to this commit, we did not check if all specified packages
were actually installed.

Setting skip_missing_names_on_install=no in yum.conf tells yum to
report errors if just one package could not be found. This revealed
an issue with the addition of the multipath and lsscsi rpms for the
Exos provider.