Luks: Fixed reserving correct size for header
Since version 2.1.0 cryptsetup by default uses luks2 for headers.
Linstor was updated to the new 16MB large headers, but unfortunately
that was done for every case, including older cryptsetups still using
luks1.

Adding the if alone was not enough, as due to the rounding effects
by i.e. LVM the resulting DRBD devices still differ in size.

Therefore also the ConfFileBuilder had to be adapted.