Mixed SPs: Allow more combinations
LVM (fat) is not allowed to skip initial sync as it does not guarantee
to zero out the device before it is access.

LVM_THIN, ZFS and ZFS_THIN are different.

That is the reason why this patch needs to blacklist all combinations
with LVM (fat) to something else but LVM (fat).

The only reason LVM_THIN was not allowed with ZFS and ZFS_THIN was
the different extent size, resulting in differnet DRBD backing
device sizes, resulting in "unrelated data".
As the previous patch fixes that issue, the 3 providers are now
compatible with each other.