BCache: Always register devices
Newer versions of bcache utils apparently already register the
just created device automatically. However, if the older version
does not do that, Linstor will wait until it runs in a timeout.

With the RetryResourceTask the next attempt will find the existing
but not registered device, so a register is called.
Unfortunately both, backing and cache device have to be registered
which means that Linstor had to run into the timeout twice.

This patch simply calls register also right after creating the
devices, that should work with new and with old versions of
bcache-utils