backup-shipping: Allow download only in used RscDfn
Before this patch the backupshippingApiHandler refused the request
if the target resource-definition had resources deployed.

With this patch the controller simply does not set the corresponding
SnapshotDefinition.Flag for restoring the snapshot into the target
ResourceDefinition after successfully downloaded.

This also allows a later "--download" flag for the client, in case
we want to implement such a thing.

This feature will also be used by L2L shipping