backup-shipping: Added cluster ID to snapshot name
Every Linstor cluster now generates via DB-migration a new local
UUID for the local cluster and stores it in controller property
"Cluster/LocalId".

This UUID is sent to the destination cluster, where the user
might have previously set the controller property
"Cluster/Remote/<uuid-from-source-cluster" to a value of his choice
(max 10 chars at the moment). This shorter description of the
source-cluster is used as a suffix for the snapshot name on the
destination cluster, resulting in LVM snapshot names like
"targetRsc_00000_back_20210625_093807_alpha"