ToggleDisk: Recreate correct layerData
DRBD_DISKLESS resources by definition only have DRBD,STORAGE layer
even if the user created the resource with an other layer stack.

This patch adds the optional Json to toggle disk such that one
can specify a custom layerstack when toggling a diskless into
a diskful. That layerstack is ignored when toggling diskful to
diskless.

If this new layerstack is not given during diskless->diskful toggle
the layerstack of the peer resources are copied, just as when
creating a new resource.

This patch also adds a check if the node actually supports all
requested / copied layers when toggling diskless into diskful.