clone-api: Introduce a way of cloning complete rscdfn + data
This adds a new clone api to the rsc-dfns, which will clone
the resource definition including all of its children and will
also copy the data of all volumes.
It will try to do this in the best way possbile, e.g.
LVM_THIN will use snapshots, ZFS will use send/recv and LVM-thick
will use a simple dd command.

As cloning can take some while depending on the size of the data,
there is also a polling API to check if a clone has already finished
or failed.