LvmThin: Changed allocated size calculation
Instead of overriding getAllocatedSize in the LvmThinProvider
this patch lets the AbsStorageProvider grad that information from
`blockdev ...`, but simply stores the actually allocated size in
the LvmThinData.

The difference is that the value returned by getAllocatedSize is
also used for setting the usable size by the AbsStorageProvider
which will result in an incorrect usable size value.

This bug went unnoticed as after a second devMgrRun the updateInfo
correctly overrides both, allocated and usable, sizes.