ToggleDisk: Fixed missing layerstack bug
Every resources has a layer-stack. The first resource of a resource-
definition stores it's layer-stack in the RD, so new resources
get the same stack unless overruled by the user.

However, if two or more resources have different layer-stacks,
the controller cannot decide which layerstack to take -> error.

In case of a toggle-disk the layerdata needs to be recreated.
That is done by deleting and rebuilding the default stack, checking
the RD for the default layer-stack and if not present the layer-stacks
of other resources.
As we just deleted the layer-stack, the resource that is toggled
will contribute an empty layer-stack, which might result in
two (or more) different stacks -> error.

To trigger this bug, the RD has to have no layer-stack stored (as
that overrules currently the layerstacks of other resources).
That is currently only possible when the resources were restored
from a snapshot