BackupShipping: Fixed ext metadata restore
Although a snapshot is created from the the external metadata
device of a given DRBD device, that snapshot is NOT shipped.

That means that when restoring from backup, there will be no
snapshot for the external metadata. Trying a usual
  linstor snapshot restore ...
from the just downloaded backup into a new resource will fail
because there is no local snapshot of the metadata device.

This patch finds this usecase and simply creates a new, empty
device for the metadata and also creates an empty snapshot from
it for consistent behaviour with the other snapshot functionality