AutoReplace: Ignore evicted resources
When a node is evicted, the autoplacer should search for nodes
to keep up the configured place-count. However, doing so the
resource of the just evicted node needs to be excluded when
counting.

Additionally changed a flux.error into a flux.just(WARN ApiCallRc)
in order to not interrupt chained fluxes.

If something breaks because of this change, please use
notConnectedError() instead of
notConnectedWarn()