BackupShipping: Do not parse time from snapName
Formating a timestamp as well as parsing it again depends on the
local timezone. That might differ from node to node.

The start timestamp is now stored as a snapDfn property and is
parsed (Long.parse instead of DateFormat.parse).

The only occurences of DateFormat.parse or .format are to interpret
user-input or render output for the user.