BackupShipping: Only collect diskfull node-ids
When restoring a backup, the metadata has to be cleared in terms of
forget-peer. However, forget-peer complains if the node-id was not
in use like as it is when using diskless peers.

Additionally, all exceptions thrown by this forget-peer command
(only during a restore operation) are ignored to not let this resource
stuck in that state.