SnapshotShipping: Fixed NPE
When a Backup is shipped (regardless if created or restored)
the SHIPPING flag is set. That flag is also used by the old
snapshot shipping.

When listing snapshot shipping, SnapshotDefinitions containing
that SHIPPING flags are filtered and properties accessed.
Running Backups will wrongly be treated as snapshot-shipments
and the missing property will cause an NPE