containers: switch to ubi8
The goal is to switch to ubi8 which was not possible because of (non
existing) package deps for s390x, but that is solved with this commit.
It also allows building on arm64 in the future.

- switch to ubi8
- use BUILDER ARG with a default that builds publicly, and allows
  setting this arg in lbbuild (to an image that already contains all the
  dependencies) without nasty sed magic. Expose that ARG in the Makefile
  as we do it for latest utils/kernel module.
- switch to newer k8s-await that is built for arm64
- avoid the public repos, but use a) the internal one (== customer
  repos), and b) official RHEL ones (we as partners are allowed to
  install/ship these packages in the image).

This makes maintainance - especially for multi arch - a lot easier, but
it reduces functionality for users that just take the Dockerfile. The
extra packages are pretty "esoteric", so I guess that should be fine.
And with Piraeus we have a good alternative anyways.