containers: pin satellite to utils 9.18.2
I am soon going to switch to RHEL8 based images after 1.15. Then we need
to provide a RHEL8 based drbd-utils image.

Pin drbd-utils to v9.18.2 which is RHEL7 based.
So old Dockerfiles will work (using latest) for now, and the pinned one
in the next release will always work.

When switching to a RHEL8 base, I do a utils release, again with
"latest" but this time based on RHEL8. This will break the then very old
Dockerfiles (expecting "latest" to be RHEL7), but not the then current
one, and not the then "old" 1.15 one which is pinned. Good enough.