PropsConEtcdDriver: Fixed removeAll method
Deleting all properties from a props-instance and adding new props
in the same transaction causes a duplicated key exception when sending
the request to ETCD.

Although this patch is quite ugly (requsting the keys to delete
from the ETCD server right before actually deleting them), but
works for now.

A more proper way to fix this would to be change the PropsConDriver
(for all kinds, sql, etcd, ...) so that the PropsContainer instance
itself is also passed to the driver. That instance can not just be
queried for the instance-name but also for all keys to delete for
this case.