RscCreate: auto toggle disk diskless->diskful
This aims to solve the scenario where we have 1 diskful and 1 diskless
resource which is not a tiebreaker and the user tries any kind of
auto-place (rd ap, rg adjust, ...) with a placecount of 2. The autoplacer
finds the storage pool on the node where the diskless resource is deployed
but was (without this patch) not brave enough to toggle-disk the
resource if it was not a tiebreaker resulting in the quite confusing
"resource already exists" message to the user.