ErrorReport: Fix compareTo method
ErrorReport has a Date field which is used in its compareTo
method. Satellites send the date as a long, which gets deserialized
into a java.util.Date instance. The controller loads the data
from its local H2 database, from where a java.sql.Timestamp (which
extends java.util.Date) is returned.

However, java.sql.Timestamp's compareTo is not symmetric with
java.util.Date's compareTo. That means that a.compareTo(b) could
be 1 while b.compareTo(a) is also 1 or 0 (but not -1 as expected)

This patch changes the returned java.sql.Timestamp to a java.sql.Date
which also extends java.util.Date, but does not override the compareTo
method, so both Date's compareTo methods are symmetric.