AbsStorageProvider: Fix control flow, style fixes
This method was supposed to add all storage pools to a list in the last statement
of a loop, but that statement was inadvertently skipped by two 'continue' statements
(which the coding standard does not allow, for exactly that reason - unless any
alternative would significantly complicate the control flow, but that was
not the case either).
Those statements were replaced by adding the 'else' branches in the two 'if'
statements that were the condition for the 'continue' statements.
Additionally, the statement that was previously skipped was moved to the top of
the loop body to prevent the bug from being reintroduced if the code section
is changed in the future.