TxSimpleObj: Fixed persistence
The k8s / CRD driver ignores the 'changed' parameter and persists
instead the entire object. However, in case of data stored through
a TransactionSimpleObject, the db.update is called before the
TxSimpleObj stored the change. Persisting the entire object via
db.update therefore takes the old version of the data.