ErrorReport: fixed missing time
The commit
5ac7e5c2e9fb0de5f41f8626968d7addc0f992c4 ErrorReport: Fix compareTo method
returned a java.sql.Date instead of java.sql.Timestamp for good reasons.
However, java.sql.Date needs to normalize the date by setting hour,
minutes and seconds to 0 to conform with the SQL 'DATE' type.

The result of this is that a 'linstor err list' showed correct dates
but all with 00:00:00 as a time. This patch should solve this issue
without re-introducing the issue fixed with 5ac7e5c2e.