docker: add "ps" to satellite container
For snapshot shipping the satellite uses "ps" to find orphaned
thin-send-recv processes. This shouldn't really matter for containers,
as we should always start in a fresh environment. The satellite currently
doesn't seem to mind if that specific command fails, on the other hand
it probably doesn't hurt to include it.