Drbd: Improve check if metadata is new
The issue this commit tries to solve is the following:
If a new resource gets created all participating satellites send
a "RequestPrimary" to the controller. Only the first request
received by the controller is answered, making the whole DRBD
resource UpToDate.

However, if the controller is restarted or due to other network
issues the "RequestPrimary" is never received by the controller
the DRBD resource will be stuck in Inconsistent state.

This commit tries to detect this scenario and resend the request.

The problem is that we must not send such a request if we are
migrating a DRBD resource from non-linstor into linstor (i.e.
checking of "DrbdPrimaryOn" on rscDfn is not enough)