Writecache,suspend: Fix check if layerstack contains DRBD
Previously the current rscLayerData was asked if DRBD is involved.
However, the current rscLayerData is only the (sub-)tree of the
actual layerData for the current layer. That means a "tree" of
DRBD->Writecache->Storage is given to the DrbdLayer to process,
but the WritecacheLayer only gets its own sub-tree, namely
Writecache->Storage which obviously does not contain DRBD anymore
making the check wrong.

The effect of this bug was "sometimes" a deadlock when trying to
resume DRBD got stuck in a D state because the writecache beneath
it was still suspended.

If all layers would suspend, the order would also be the opposite
when resuming. IF we want to suspend on multiple layers, the suspend
needs to be top-down (DRBD,Writecache,...) but the resume needs
to be bottom-up (...,Writecache,DRBD).