clone: allow a new zfs-clone mode with zfs storage
This mode can either be enabled by a rest parameter or
a controller/rscgrp property.
It will use `zfs clone` instead of zfs send/recv and should be way faster,
with the downside of having a parent-child relation between the clones.
This doesn't matter in a cloudstack environment and might be also used for
opennebula setups.