backups: allow customizing the snapshot name
Some plugins (namely CSI) basically require control over the snapshot
name. This was a problem when starting to integrate backups, as there the
snapshot name was not deterministic, and so there was no way to identify
which backup operation was associated with plugin request.

To solve this, this commit introduces a way to:
1. set the snapshot name for the snapshot on the local cluster when requesting
   a backup.
2. encode that snapshot name in the S3 object.

With these changes a plugin can use it's own ID for the snapshot name, and
can later use it when selecting the right backup object to restore from.

To maintain backwards-compatibility, the snapshot name is encoded as part of
the existing S3-suffix. As a consequence of the encoding, the S3-suffix can no
longer contain `^` without breaking the encoding. This is probably acceptable,
as nobody knows about this feature anyways.