CRD,Flags: Fixed updating flags
When updating a flag, ETCD and SQL use the passed parameter for the
new flags to perform / prepare the update.
CRD persists the whole object, so it ignored the passed parameter.

The StateBitsFlag first called the .persist and just afterwards
updated the Linstor data, which leads to persisting the old flags
instead of the new.

Swapping those lines fixes the flag-updating.

Unfortunately swapping those lines also breaks the nice logging
of ETCD and SQL, which use the data's value (old flags) to
print the fromBits and the parameter to print the toBits.
Swapping those lines makes both variables the same. Therefore
the rest of this patch introduces a new variable for the flag-
persistance so ETCD and SQL can refer to the old value.