Crd: have separate create/replace operations
We can reduce the number of API calls we have to do based on our
knowledge of the resource state. For example, if we know that we
update an existing resource, we don't have to execute an
(always failing) create operation first.

This also means we run into errors quicker should the cluster state
get ever out of sync with the data store.

Instead of having a single update method on the CRD transaction object,
there is now two methods: create & replace. Note that "update" isn't
available any longer. This prevents any dangling reference to the
update method, still expecting the old "upsert" behaviour.