Crd: ensure rollback is in known state
There should only ever be one rollback resource defined. To ensure
this is the case, store the state of the rollback resource in the
transaction.

We only load the rollback resource from the DB at the start of the migration.
This should ensure that the DB recovers to a known state after a controller
restart during a commit.