toggle-disk: fix inconsistent DB state
Toggling from diskfull to diskless had two bugs that mostly cancelled
each other out, at least when using SQL or ETCD DB drivers. Since the
K8s DB driver is more strict in what it accepts, toggle-disk and
all operations using it (node evacuate, deleting a resource that is
kept as tie-breaker) failed, leaving the DB in a bad state.

The first issue is related to how a changing storage pool was handled:
Replacement should be a simple "delete old + add new" operation. However,
due to the way the implementation works, it instead deleted the old data
from the DB while keeping it in the internal state. This caused the DB
to miss an entry for the storage volume for a short time.

The second issue was that the layer stack was updated unconditionally
with the supposed "new" layers, before the satellite was informed about
the requested removal of the storage layer.

These errors mostly cancelled each other out, since when the storage layer
removal was actually executed by the satellite, the controller reported the
"updated" internal state, which is actually still the old state. A later
stage then rebuilds the layer stack again, this time first clearing out
the old data, preventing the same mistake happening twice.

This mostly worked for SQL and ETCD drivers, as they don't do any sanity
checks on reloading the layers from the DB. Since the Controller will
trigger a retry of the disk remove operation immediately after the
restart, the missing DB entry is created shortly afterwards. The K8s
driver meanwhile has such a sanity check.