DevMgr,Flux: Add response for noop
In case the satellite requests i.e. a resource and the controller
responses with "ApplyDeleted*" (which means the resource is already
deleted since the controller told the satellite "ChangedRsc") the
satellite will obviously perform a noop or max some cleanups.
The noop will NOT invoke the DeviceHandler. However, only the
DeviceHandler calls the corresponding respond method which tells
the controller that the resource was successfully "applied".

Omitting that message triggers the satellite-error "No response
generated by handler" since there are indeed no messages
in the flux queue before calling its .complete()