autoplacer: account for over-provisioning
The AutoPlacer now takes the overprovisioning setting into account when
calculating the free capacity on a storage pool. That means it is possible
to create volumes that are larger than the remaining, physically free capacity.

That also means that we might run into situations where DRBD starts syncing to
nodes that are already at capacity limit, causing out-of-space errors. Let's
call those situations "user errors" for now.

In theory you can guess-timate the actually used size of a volume based on
observing the current state (new volume -> expect a few kb for DRBD, existing
volume -> expect biggest allocation on peers). This was deemed to
complicated for the current implementation.