docker: clean keystore before import
keytool doesn't like to import the same key twice in the same keystore.
There are a few situations in which the keystore might survive the container
restart, so we have to ensure the import runs along even if the old keystore
is still present. The simple solution is to just remove the keystore on import.