TaskScheduler: Add reschedule method
Sometimes a task needs to be rescheduled or even canceled due to
external actions (i.e. through an apiCallHandler that changes
a "run next" setting of an existing task).
This new method requeues the task in the internal map of next
tasks.

Using this new method, we could rework some of our task-implementations
to create smaller, but more tasks instead of having one task that
handles sub-tasks (i.e. ReconnectorTask, AutoSnapshotTask, etc..)