BackupShipping: Use correct flux order
The flux step that removes a snapshot from the "shipping in progress"
list needs to be executed after the stlt updates the snap-flag that
can trigger a shipping. Otherwise the shipping might be started twice
if the timing is just right which will lead to unwanted effects and/or
errors.