TaskSchedule: Added new next execution-time-calculation
The previous version had the issue that if a task always returns
i.e. 10s after the scheduled execution, if the execution was delayed
for quite a long time (i.e. 30 sec), the scheduledAt would still be
in the past, where +10s would still be 20 sec ago, which means
that task would be immediately rescheduled 3 times in a row.

The new calulations calculates the next timestamp that is a multiple
of the relative (10s) away starting from the scheduledAt.